requirements.txt
web3==6.0.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
pandas-ta==0.3.14b0
tensorflow==2.13.0
tensorboardX==2.6.1
